requirements.txt
firebase-admin
pandas
google-generativeai